test: Update test code
