+++ backend/main-service/src/test/java/io/blog/devlog/domain/post/repository/PostRepositoryTest.java
@@ -23,7 +23,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 @DataJpaTest
 @ActiveProfiles("test")
 @AutoConfigureTestDatabase(connection = EmbeddedDatabaseConnection.H2)
-@Sql(value= "/testsql/post/post.sql", executionPhase = Sql.ExecutionPhase.BEFORE_TEST_CLASS)
+@Sql(value="/testsql/post/post.sql", executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
 public class PostRepositoryTest {
     @Autowired
     private PostRepository postRepository;
